Revert "refactor(desktop): lift the startup route load and screen gate out of App"
This reverts b97c2a8. The extraction put multi-source orchestration behind a UI hook, which AGENTS.md rule 5 reserves for a service, so it needs a shape this PR is not the place to settle. The connectivity fix stands on its own without it.

App is back over react-doctor's 300-line no-giant-component threshold, which is advisory and does not fail the job. Splitting App belongs in its own change.

Generated-By: PostHog Desktop
Task-Id: a8e8600f-93fd-40f6-a48f-06b546a38844
